move code for Go modules support to gomod.go
