+++ session.go
@@ -1,7 +1,6 @@
 package gore
 
 import (
-	"bufio"
 	"bytes"
 	"errors"
 	"fmt"
@@ -18,7 +17,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"syscall"
 	"unicode"
@@ -91,57 +89,6 @@ func NewSession(stdout, stderr io.Writer) (*Session, error) {
 	}
 
 	return s, nil
-}
-
-func (s *Session) initGoMod() (err error) {
-	hasMod, replaces, err := getModReplaces()
-	if err != nil || !hasMod {
-		return
-	}
-
-	tempModule := filepath.Base(s.tempDir)
-	goModPath := filepath.Join(s.tempDir, "go.mod")
-
-	mod := "module " + tempModule + "\n" + strings.Join(replaces, "\n")
-
-	return ioutil.WriteFile(goModPath, []byte(mod), 0644)
-}
-
-func getModReplaces() (hasMod bool, replaces []string, err error) {
-	pwd, err := os.Getwd()
-	if err != nil {
-		return
-	}
-
-	file, err := os.Open(filepath.Join(pwd, "go.mod"))
-	if err != nil {
-		if os.IsNotExist(err) {
-			err = nil
-		}
-		return
-	}
-	defer file.Close()
-
-	out, err := exec.Command("go", "list", "-m", "all").Output()
-	s := bufio.NewScanner(bytes.NewReader(out))
-
-	s.Scan()
-	module := s.Text()
-	if module == "" {
-		return
-	}
-
-	hasMod = true
-	replaces = append(replaces, "replace "+module+" => "+strconv.Quote(pwd))
-
-	for s.Scan() {
-		replace := s.Text()
-		if strings.Contains(replace, "=>") {
-			replaces = append(replaces, "replace "+replace)
-		}
-	}
-
-	return
 }
 
 func (s *Session) init() (err error) {
